refactor proxy
